requirements.txt
streamlit
tensorflow
pillow
requests
python-multipart
